scripts/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.0
Pillow==10.0.0
facenet-pytorch==2.5.2
torch==2.0.1
torchvision==0.15.2